move service test in services
