+++ src/Services/ResolveLocation/Resolver/ResolverInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Seatplus\Eveapi\Services\ResolveLocation\Resolver;
+
+use Exception;
+use Seatplus\Eveapi\Models\Universe\Location;
+
+interface ResolverInterface
+{
+    /**
+     * @throws Exception
+     */
+    public function handle(Location $location): bool;
+}
