+++ tests/Unit/Services/ResolveLocation/StructureResolverTest.php
@@ -6,9 +6,9 @@ use Seatplus\Eveapi\Models\RefreshToken;
 use Seatplus\Eveapi\Models\Universe\Location;
 use Seatplus\Eveapi\Models\Universe\Station;
 use Seatplus\Eveapi\Models\Universe\Structure;
-use Seatplus\Eveapi\Services\ResolveLocation\StructureRefreshTokenFinder;
-use Seatplus\Eveapi\Services\ResolveLocation\StructureResolver;
+use Seatplus\Eveapi\Services\ResolveLocation\Resolver\StructureResolver;
 use Illuminate\Support\Facades\Queue;
+use Seatplus\Eveapi\Services\ResolveLocation\StructureRefreshTokenFinder;
 
 beforeEach(function () {
     Queue::fake();
